Most of 4.5
All the functionalities of 4.5 added, just need to fix some bugs.

Able to filter by city, change the number of events wanted and have their corresponding integration and unit tests created.

Currently a bug is occurring where an error is popping up with the hook for NumberOfEvents. Will fix in next the version.